requirements.txt
Django>=2.0,<3.0
requests==2.20.0
mongoengine==0.16
django-allauth==0.39.1
djangorestframework==3.9.3
django-rest-framework-mongoengine==3.3.1
